+++ js/util.js
@@ -1,0 +1,16 @@
+
+const getRandomPositiveInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const checkStringLength = (string , length) => string.length <= length;
+
+const getRandomArrayElement = (array) =>
+  array[getRandomPositiveInteger(0, array.length - 1)];
+
+export {getRandomArrayElement};
+export {checkStringLength};
+export {getRandomPositiveInteger};
